Add basic types and trait implementations for fs::Tree
I switched the tree back to cloning its items. This allows users of thetree to wrap items in an Arc if they want to, which makes it easier toutilize the fact that items are heap allocated elsewhere in the system.In this case, I wanted to avoid cloning entries in order to place themwithin relative paths.I use a SmallVec of size 1 because the relative paths summarizingindividual items will always be 1 element long, which makes a heapallocation in that case wasteful. I don't want to reserve space for morethan one element however, because that would end up bloating the size ofthe tree.