Get randomized tests passing with 1 tree update
I've changed the interface of the tree builder slightly so that we can
detect deleted entries at the end of the targeted root directory. Now
you're required to push and pop the directory at the path passed to the
builder on construction. When the root is popped, it allows us to detect
any entries that were removed from the end of the root.