Refactor fs::Builder
This commit simplifies the Builder code and sets us up to traverse 
across moved directories by eliminating direct interaction with the 
cursor in Builder::push and Builder::tree. These methods now interact 
with helper methods that call the cursor, which paves the way for the 
builder to maintain a *stack* of cursors into the old tree which allow 
it to traverse moved directories as we push entries during a new file 
system scan.